Decouple microwave oven control cluster part 1
