+++ src/app/clusters/microwave-oven-control-server/app_config_dependent_sources.cmake
@@ -1,4 +1,4 @@
-# Copyright (c) 2025 Project CHIP Authors
+# Copyright (c) 2025-2026 Project CHIP Authors
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -16,6 +16,8 @@
 TARGET_SOURCES(
   ${APP_TARGET}
   PRIVATE
-    "${CLUSTER_DIR}/microwave-oven-control-server.cpp"
+    "${CLUSTER_DIR}/MicrowaveOvenControlCluster.cpp"
+    "${CLUSTER_DIR}/CodegenIntegration.h"
     "${CLUSTER_DIR}/microwave-oven-control-server.h"
+    "${CLUSTER_DIR}/Delegate.h"
 )
