+++ src/app/clusters/microwave-oven-control-server/app_config_dependent_sources.gni
@@ -1,4 +1,4 @@
-# Copyright (c) 2025 Project CHIP Authors
+# Copyright (c) 2025-2026 Project CHIP Authors
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,6 +12,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 app_config_dependent_sources = [
-  "microwave-oven-control-server.cpp",
+  "MicrowaveOvenControlCluster.cpp",
+  "CodegenIntegration.h",
   "microwave-oven-control-server.h",
+  "Delegate.h",
 ]
